axiosUtils removed, sitemapUtils and urlUtils removed and migrated to new sitemapHandler, also now including XML file results
