--- src/utils/__mocks__/sitemapIndex.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<sitemapindex xmlns="http://www.google.com/schemas/sitemap/0.84">
-    <sitemap>
-        <loc>https://www.google.com/gmail/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/forms/sitemaps.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/slides/sitemaps.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/sheets/sitemaps.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/drive/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/docs/sitemaps.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/get/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/flights/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/admob/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/business/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/services/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/partners/about/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/adwords/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/search/about/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/adsense/start/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/retail/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/sitemap_search.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/webmasters/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/chromebook/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/chrome/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/calendar/about/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/photos/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/nonprofits/sitemap.xml</loc>
-    </sitemap>
-    <sitemap>
-        <loc>https://www.google.com/finance/sitemap.xml</loc>
-    </sitemap>
-</sitemapindex>
