+++ src/utils/index.ts
@@ -1,4 +1,1 @@
-export * from './axiosUtils';
 export * from './envUtils';
-export * from './urlUtils';
-export * from './sitemapUtils';
